resolve the recycle import problecm

fix issue


fix ci
